Add function to get token pool price
